COM-2174: add qrcode button to helper
