+++ src/modules/client/components/customers/infos/ProfileInfo.vue
@@ -186,7 +186,7 @@
     <div class="q-mb-lg">
       <p class="text-weight-bold">Justificatifs APA ou autres financements</p>
       <div class="row gutter-profile q-mb-lg">
-        <div class="col-xs-12">
+        <div class="col-xs-12 col-md-6">
           <ni-multiple-files-uploader path="financialCertificates" alt="justificatif financement" drive-storage
             @uploaded="documentUploaded" name="financialCertificates" collapsible-label="Ajouter un justificatif"
             :user-profile="customer" :url="docsUploadUrl" @delete="validateFinancialCertifDeletion($event)"
