+++ src/modules/client/pages/customers/Subscriptions.vue
@@ -39,14 +39,18 @@
           </div>
         </div>
       </div>
-      <div class="q-mb-lg">
+      <div class="q-mb-md">
         <p class="title">Justificatifs APA ou autres financements</p>
-        <div class="row gutter-profile">
-          <div class="col-xs-12">
+        <div class="row items-center">
+          <div class="col-xs-12 col-md-6 q-mr-lg">
             <ni-multiple-files-uploader path="financialCertificates" alt="justificatif_financement" :url="docsUploadUrl"
               @uploaded="documentUploadedForFinancialCertificates" name="financialCertificates" drive-storage
               collapsible-label="Ajouter un justificatif" :user-profile="customer" :extensions="extensions" multiple
               @delete="validateFinancialCertifDeletion($event)" additional-fields-name="justificatif_financement" />
+          </div>
+          <div class="q-mb-md">
+            <p class="input-caption">Horodatage</p>
+            <ni-bi-color-button icon="file_download" label="QR Code" size="16px" @click="downloadQRCode()" />
           </div>
         </div>
       </div>
@@ -137,6 +141,7 @@ import TitleHeader from '@components/TitleHeader';
 import Input from '@components/form/Input';
 import MultipleFilesUploader from '@components/form/MultipleFilesUploader';
 import Button from '@components/Button';
+import BiColorButton from '@components/BiColorButton';
 import Modal from '@components/modal/Modal';
 import HtmlModal from '@components/modal/HtmlModal';
 import ResponsiveTable from '@components/table/ResponsiveTable';
@@ -164,6 +169,7 @@ export default {
     'ni-html-modal': HtmlModal,
     'ni-responsive-table': ResponsiveTable,
     'ni-funding-grid-table': FundingGridTable,
+    'ni-bi-color-button': BiColorButton,
   },
   mixins: [customerMixin, subscriptionMixin, financialCertificatesMixin, fundingMixin, tableMixin],
   data () {
